Modification of the manifest abstract model (proposal)
It only conflates requirements 3 and 8 in 3, because the sequence of primary resources will defines the default reading order, whatever concrete syntax is chosen. It also adds a note about the Web Publication address, which references the requirements this property fulfills (I think that in the current editor's version, we should reference requirements and maybe create some new ones).